Changes to startup
